feat: improve test setup
