+++ playwright.config.js
@@ -1,6 +1,10 @@
 // playwright.config.js
 const { defineConfig, devices } = require("@playwright/test");
 
+const ADDON_NAME = process.env.ADDON_NAME || "tourGuide";
+if (!process.env.ADDON_NAME) {
+  console.warn("ADDON_NAME environment variable not set, using default: exporter");
+}
 const BASE_URL = process.env.BASE_URL || "http://localhost:8080";
 
 // Optional override for custom Chromium binary
@@ -8,8 +12,11 @@ const CHROMIUM_EXECUTABLE =
     process.env.PLAYWRIGHT_CHROMIUM_EXECUTABLE_PATH || undefined;
 
 module.exports = defineConfig({
-    testDir: ".",
-    testMatch: ["tests/e2e/*.spec.js"],
+    testDir: "./test/e2e",
+    testMatch: [
+        "addon.spec.js",
+        `**/${ADDON_NAME}.spec.js`,
+    ],
     retries: process.env.CI ? 1 : 0,
     workers: 1,
     reporter: [
