+++ test/e2e/config/resources/newsFeedPortalAlerts.json
