+++ test/e2e/config/resources/rest-services.json
