+++ test/e2e/config/resources/sql-wasm.wasm
