+++ test/e2e/config/resources/style_v3.json
